# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dondok API
  version: 0.10.0
  description: 돈독 프론트엔드와 백엔드 사이의 HTTP 계약
  license:
    name: All rights reserved
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security: []
paths:
...
  /api/transactions/calendar:
    get:
      summary: 월 범위 날짜별 수입·지출 합계
      operationId: getTransactionCalendar
      tags: [Transaction]
      security: [{ cookieAuth: [] }]
      parameters:
        - name: month
          in: query
          required: true
          schema: { type: string, pattern: "^[0-9]{4}-[0-9]{2}$" }
      responses:
        "200":
          description: 이체와 조정을 제외한 월·일 합계
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransactionCalendar" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/transactions:
    get:
      summary: 기간 내 거래 cursor 목록
      operationId: getTransactions
      tags: [Transaction]
      security: [{ cookieAuth: [] }]
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: toExclusive
          in: query
          required: true
          schema: { type: string, format: date }
        - name: cursor
          in: query
          required: false
          description: 서버가 발급한 날짜·생성 시각·ID 기반 opaque cursor
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: 최신 날짜 우선 거래 목록
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransactionPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: 수입·지출·일반 이체와 posting 생성
      operationId: createTransaction
      tags: [Transaction]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/CsrfHeader" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - { $ref: "#/components/schemas/CreateIncomeRequest" }
                - { $ref: "#/components/schemas/CreateExpenseRequest" }
                - { $ref: "#/components/schemas/CreateTransferRequest" }
              discriminator:
                propertyName: type
      responses:
        "201":
          description: 생성된 거래·posting·카드 회차의 authoritative 상태
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/transactions/{transactionId}:
    parameters:
      - name: transactionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 현재 가계부의 거래 상세
      operationId: getTransaction
      tags: [Transaction]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: 거래·posting과 일반 관리 가능 유형
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      summary: expected version 기준 일반 거래 전체 수정
      description: 거래 종류는 변경할 수 없고 신용카드 구매와 시스템 거래는 별도 command를 사용한다.
      operationId: updateTransaction
      tags: [Transaction]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTransactionRequest" }
      responses:
        "200":
          description: 변경된 거래와 posting의 authoritative 상태
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
    delete:
      summary: expected version 기준 일반 거래 soft delete
      operationId: deleteTransaction
      tags: [Transaction]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/CsrfHeader" }
        - name: expectedVersion
          in: query
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
      responses:
        "200":
          description: 삭제된 거래 식별자와 새 version
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletedTransaction" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
  /api/transactions/{transactionId}/card-purchase-management:
    parameters:
      - name: transactionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 카드 구매의 명세·결제·환불 관리 상태 조회
      operationId: getCardPurchaseManagement
      tags: [Transaction]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: 원 구매와 현재 환불 가능액을 포함한 authoritative 관리 상태
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardPurchaseManagement" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/transactions/{transactionId}/card-purchase-corrections/preview:
    parameters:
      - name: transactionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: 카드 구매 기록 정정의 명세·결제 계좌 영향 확인
      operationId: previewCardPurchaseCorrection
      tags: [Transaction]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CardPurchaseCorrectionRequest" }
      responses:
        "200":
          description: 저장 직전 사용자에게 보여줄 authoritative 영향
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardPurchaseCorrectionPreview" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
  /api/transactions/{transactionId}/card-purchase-corrections:
    parameters:
      - name: transactionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: 카드 구매 기록과 명세·정산을 원 구매일 기준으로 소급 정정
      operationId: applyCardPurchaseCorrection
      tags: [Transaction]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/CsrfHeader" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplyCardPurchaseCorrectionRequest" }
      responses:
        "200":
          description: 정정된 구매와 결제·환불 관리 상태
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardPurchaseManagement" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
  /api/transactions/{transactionId}/card-purchase-refunds/preview:
    parameters:
      - name: transactionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: 실제 카드 구매 환불의 미결제분·원 결제 계좌 반환 영향 확인
      operationId: previewCardPurchaseRefund
      tags: [Transaction]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CardPurchaseRefundRequest" }
      responses:
        "200":
          description: 저장 직전 사용자에게 보여줄 authoritative 환불 영향
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardPurchaseRefundPreview" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
  /api/transactions/{transactionId}/card-purchase-refunds:
    parameters:
      - name: transactionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: 실제 환불일의 지출 상쇄와 카드·원 결제 계좌 반환 기록
      operationId: applyCardPurchaseRefund
      tags: [Transaction]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/CsrfHeader" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplyCardPurchaseRefundRequest" }
      responses:
        "201":
          description: 생성된 환불 거래와 카드·계좌 반영 결과
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardPurchaseRefundResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: DONDOK_SESSION
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 같은 사용자의 동일 키·동일 payload 재시도는 기존 리소스의 동일 명령 결과를 반환한다.
      schema: { type: string, minLength: 1, maxLength: 100 }
  schemas:
...
    TransactionType:
      type: string
      enum: [INCOME, EXPENSE, TRANSFER]
    TransferSubtype:
      type: string
      enum: [NORMAL, CARD_SETTLEMENT, CARD_PREPAYMENT]
    TransactionInput:
      type: object
      required: [type, occurredOn, amountWon, performedByMemberId]
      properties:
        type: { $ref: "#/components/schemas/TransactionType" }
        occurredOn: { type: string, format: date }
        amountWon: { type: integer, format: int64, minimum: 1 }
        performedByMemberId: { type: string, format: uuid }
        description: { type: string, maxLength: 500 }
    CreateIncomeRequest:
      allOf:
        - { $ref: "#/components/schemas/TransactionInput" }
        - type: object
          required: [categoryId, assetId]
          properties:
            type: { type: string, const: INCOME }
            categoryId: { type: string, format: uuid }
            assetId: { type: string, format: uuid }
    CreateExpenseRequest:
      allOf:
        - { $ref: "#/components/schemas/TransactionInput" }
        - type: object
          required: [categoryId, assetId]
          properties:
            type: { type: string, const: EXPENSE }
            categoryId: { type: string, format: uuid }
            assetId: { type: string, format: uuid }
            installmentCount: { type: integer, minimum: 1, maximum: 60, default: 1 }
    CreateTransferRequest:
      allOf:
        - { $ref: "#/components/schemas/TransactionInput" }
        - type: object
          required: [sourceAssetId, destinationAssetId]
          properties:
            type: { type: string, const: TRANSFER }
            sourceAssetId: { type: string, format: uuid, description: "같은 가계부의 활성 BANK 계좌 자산 ID" }
            destinationAssetId: { type: string, format: uuid, description: "출발 계좌와 다른 활성 BANK 계좌 자산 ID" }
    UpdateTransactionRequest:
      type: object
      required: [type, occurredOn, amountWon, performedByMemberId, expectedVersion]
      properties:
        type: { $ref: "#/components/schemas/TransactionType" }
        occurredOn: { type: string, format: date }
        amountWon: { type: integer, format: int64, minimum: 1 }
        categoryId: { type: string, format: uuid }
        assetId: { type: string, format: uuid }
        sourceAssetId: { type: string, format: uuid, description: "TRANSFER 수정에서만 사용하는 활성 BANK 계좌 자산 ID" }
        destinationAssetId: { type: string, format: uuid, description: "TRANSFER 수정에서만 사용하는 출발과 다른 활성 BANK 계좌 자산 ID" }
        performedByMemberId: { type: string, format: uuid }
        description: { type: string, maxLength: 500 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    TransactionCategory:
      type: object
      required: [categoryId, name]
      properties:
        categoryId: { type: string, format: uuid }
        name: { type: string }
...
    Transaction:
      type: object
      required:
        - transactionId
        - type
        - transferSubtype
        - managementType
        - relatedPurchaseTransactionId
        - occurredOn
        - amountWon
        - category
        - performedBy
        - createdBy
        - asset
        - description
        - postings
        - installmentCount
        - version
        - createdAt
        - updatedAt
      properties:
        transactionId: { type: string, format: uuid }
        type: { $ref: "#/components/schemas/TransactionType" }
        transferSubtype:
          oneOf:
            - { $ref: "#/components/schemas/TransferSubtype" }
            - { type: "null" }
        managementType:
          type: string
          enum: [GENERAL, CARD_PURCHASE, CARD_REFUND, SYSTEM]
        relatedPurchaseTransactionId:
          description: 카드 환불 거래가 연결된 원 구매 식별자이며 그 외에는 null이다.
          oneOf:
            - { type: string, format: uuid }
            - { type: "null" }
        occurredOn: { type: string, format: date }
        amountWon: { type: integer, format: int64 }
        category:
          oneOf:
            - { $ref: "#/components/schemas/TransactionCategory" }
            - { type: "null" }
        performedBy:
          oneOf:
            - { $ref: "#/components/schemas/TransactionMember" }
            - { type: "null" }
        createdBy:
          oneOf:
            - { $ref: "#/components/schemas/TransactionMember" }
            - { type: "null" }
        asset:
          description: 수입·지출에서 사용자가 선택한 자산이다. 체크카드는 실제 posting 계좌와 다를 수 있다.
          oneOf:
            - { $ref: "#/components/schemas/TransactionAsset" }
            - { type: "null" }
        description:
          oneOf:
            - { type: string }
            - { type: "null" }
        postings:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/TransactionPosting" }
        installmentCount:
          oneOf:
            - { type: integer, minimum: 1 }
            - { type: "null" }
        version: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CardPurchaseManagement:
      type: object
      required: [purchase, billingSnapshot, refundableAmountWon, charges, statements, refunds]
      properties:
        purchase: { $ref: "#/components/schemas/Transaction" }
        billingSnapshot: { $ref: "#/components/schemas/CardPurchaseBillingSnapshot" }
        refundableAmountWon: { type: integer, format: int64, minimum: 0 }
        charges:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/CardPurchaseCharge" }
        statements:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/CardPurchaseStatement" }
        refunds:
          type: array
          items: { $ref: "#/components/schemas/CardPurchaseRefund" }
    CardPurchaseCorrectionRequest:
      type: object
      required: [occurredOn, amountWon, categoryId, cardAssetId, performedByMemberId, installmentCount, expectedVersion]
      properties:
        occurredOn: { type: string, format: date }
        amountWon: { type: integer, format: int64, minimum: 1 }
        categoryId: { type: string, format: uuid }
        cardAssetId: { type: string, format: uuid }
        performedByMemberId: { type: string, format: uuid }
        description: { type: string, maxLength: 500 }
        installmentCount: { type: integer, minimum: 1, maximum: 60 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    ApplyCardPurchaseCorrectionRequest:
      allOf:
        - { $ref: "#/components/schemas/CardPurchaseCorrectionRequest" }
        - type: object
          required: [previewToken]
          properties:
            previewToken: { type: string, minLength: 1, maxLength: 128 }
    CardPurchaseCorrectionPreview:
      type: object
      required: [previewToken, purchaseVersion, unpaidCardReductionWon, accountReturns]
      properties:
        previewToken: { type: string }
        purchaseVersion: { type: integer, format: int64, minimum: 0 }
        unpaidCardReductionWon: { type: integer, format: int64, minimum: 0 }
        accountReturns:
          type: array
          items: { $ref: "#/components/schemas/CardPurchaseAccountReturn" }
    CardPurchaseRefundRequest:
      type: object
      required: [refundedOn, amountWon, expectedVersion]
      properties:
        refundedOn: { type: string, format: date }
        amountWon: { type: integer, format: int64, minimum: 1 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        description: { type: string, maxLength: 500 }
    ApplyCardPurchaseRefundRequest:
      allOf:
        - { $ref: "#/components/schemas/CardPurchaseRefundRequest" }
        - type: object
          required: [previewToken]
          properties:
            previewToken: { type: string, minLength: 1, maxLength: 128 }
    CardPurchaseRefundPreview:
      type: object
      required: [previewToken, purchaseVersion, refundableAmountWon, unpaidCardReductionWon, accountReturns]
      properties:
        previewToken: { type: string }
        purchaseVersion: { type: integer, format: int64, minimum: 0 }
        refundableAmountWon: { type: integer, format: int64, minimum: 0 }
        unpaidCardReductionWon: { type: integer, format: int64, minimum: 0 }
        accountReturns:
          type: array
          items: { $ref: "#/components/schemas/CardPurchaseAccountReturn" }
    CardPurchaseRefundResult:
      type: object
      required: [purchase, refundTransaction, unpaidCardReductionWon, accountReturns]
      properties:
        purchase: { $ref: "#/components/schemas/Transaction" }
        refundTransaction: { $ref: "#/components/schemas/Transaction" }
        unpaidCardReductionWon: { type: integer, format: int64, minimum: 0 }
        accountReturns:
          type: array
          items: { $ref: "#/components/schemas/CardPurchaseAccountReturn" }
    DeletedTransaction:
      type: object
      required: [transactionId, deletedVersion]
      properties:
        transactionId: { type: string, format: uuid }
        deletedVersion: { type: integer, format: int64, minimum: 1 }
    TransactionDaySummary:
      type: object
      required: [date, incomeWon, expenseWon, netWon]
      properties:
        date: { type: string, format: date }
        incomeWon: { type: integer, format: int64 }
        expenseWon: { type: integer, format: int64 }
        netWon: { type: integer, format: int64 }
    TransactionCalendar:
      type: object
      required: [month, totalIncomeWon, totalExpenseWon, netWon, days]
      properties:
        month: { type: string, pattern: "^[0-9]{4}-[0-9]{2}$" }
        totalIncomeWon: { type: integer, format: int64 }
        totalExpenseWon: { type: integer, format: int64 }
        netWon: { type: integer, format: int64 }
        days:
          type: array
          items: { $ref: "#/components/schemas/TransactionDaySummary" }
...
    TransactionPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Transaction" }
        nextCursor:
          oneOf:
            - { type: string }
            - { type: "null" }
    Problem:
      type: object
      required: [status, title, detail, errorCode, correlationId, timestamp, fieldErrors]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        errorCode: { type: string }
        correlationId: { type: string }
        timestamp: { type: string, format: date-time }
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: true
  responses:
    BadRequest:
      description: 입력값 또는 token 오류
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Unauthorized:
      description: 인증 실패
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
    Conflict:
      description: 현재 상태와 충돌하는 요청
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    NotFound:
      description: 현재 사용자가 접근할 수 있는 리소스가 없음
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    PreconditionFailed:
      description: 편집 시작 후 다른 세션이 리소스를 수정함
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }